requirements.txt
numpy==1.24.3
pandas==2.0.3
streamlit==1.28.0
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.6.1
PyPDF2==3.0.1
python-docx==0.8.11
plotly==5.15.0
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2
Pillow==10.0.0
typing-extensions==4.7.1
protobuf==3.20.3